@@ -1,7 +1,7 @@
 {
   "name": "triggercmdagent",
   "productName": "TRIGGERcmdAgent",
-  "version": "1.0.25",
+  "version": "1.0.26",
   "description": "TRIGGERcmd Windows Agent",
   "main": "src/main.js",
   "scripts": {
@@ -50,11 +50,13 @@
     "bootstrap": "^3.3.7",
     "child_process": "^1.0.2",
     "chokidar": "^1.6.1",
-    "electron-compile": "^5.1.0",
+    "electron-compile": "^6.4.4",
     "electron-reload": "^1.0.2",
     "electron-squirrel-startup": "^1.0.0",
     "fs": "0.0.1-security",
-    "jquery": "^3.1.0",
+    "i18next": "^21.9.1",
+    "i18next-browser-languagedetector": "^6.1.5",
+    "jquery": "^3.6.1",
     "lodash": "^4.13.1",
     "minimist": "^1.2.0",
     "node-windows": "^0.1.12",
@@ -62,14 +64,26 @@
     "path": "^0.12.7",
     "posix-argv-parser": "^2.0.0",
     "prompt": "^1.0.0",
+    "react-i18next": "^11.18.6",
     "regedit": "^2.2.6",
     "request": "^2.79.0",
     "sails.io.js": "^1.1.4",
     "socket.io-client": "^2.0.2"
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.19.1",
+    "@babel/preset-react": "^7.18.6",
+    "babel-core": "7.0.0-bridge.0",
+    "babel-loader": "^8.2.5",
+    "babel-plugin-transform-jquery": "^1.0.3",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.16.0",
-    "electron-prebuilt-compile": "1.4.12",
+    "babelify": "^10.0.0",
+    "browserify": "^17.0.0",
+    "browserify-shim": "^3.8.15",
+    "electron-prebuilt-compile": "8.2.0",
     "electron-rebuild": "^1.5.6",
     "eslint": "^3.12.2",
     "eslint-config-airbnb": "^13.0.0",
@@ -77,12 +91,17 @@
     "eslint-plugin-jsx-a11y": "^2.2.3",
     "eslint-plugin-react": "^6.8.0",
     "gulp": "^4.0.2",
+    "gulp-babel": "^6.1.3",
     "gulp-browserify": "^0.5.1",
     "gulp-concat-css": "^2.3.0",
-    "gulp-react": "^3.1.0",
+    "gulp-reactify": "^3.0.1",
     "gulp-run": "^1.7.1",
-    "react": "^15.1.0",
-    "react-dom": "^15.1.0",
-    "reactify": "^1.1.1"
+    "gulp-uglify": "^3.0.2",
+    "gulplog": "^2.0.0",
+    "react": "^16.8.0",
+    "react-dom": "^16.8.0",
+    "reactify": "^1.1.1",
+    "vinyl-buffer": "^1.0.1",
+    "vinyl-source-stream": "^2.0.0"
   }
 }
