@@ -1,7 +1,79 @@
-var $ = jQuery = require('jquery');
+import i18next from "i18next";
+import { initReactI18next } from "react-i18next";
+var fs = eRequire('fs');
+// import LanguageDetector from 'i18next-browser-languagedetector';
+
+var lang;
+try {
+  lang = fs.readFileSync(languageLocation).toString();;
+  console.log("Found " + lang + " in " + languageLocation);
+} catch (err) {
+  console.log("No language found in " + languageLocation + " using en.");
+  lang = 'en';
+}
+
+i18next
+  .use(initReactI18next)
+  .init({
+    lng: lang,
+    fallbackLng: 'en',
+    debug: false,
+    react: {
+      useSuspense: true
+    },
+    resources: {
+      en: {
+        translation: {
+          "Add Command": "Add Command",
+          "Browse Examples": "Browse Examples",
+          "Computer List": "Computer List",
+          "Operating System": "Operating System",
+          "Cancel": "Cancel",
+          "Allow Parameters": "Allow Parameters",
+          "Save": "Save",
+          "Alexa or Google will say this when it runs (optional)": "Alexa or Google will say this when it runs (optional)",
+          "How to use background commands": "How to use background commands",
+          "How to use Off Command": "How to use Off Command",
+          "Voice Reply": "Voice Reply",
+          "Word you\'ll say to Alexa or Google (optional)": "Word you\'ll say to Alexa or Google (optional)",
+          "If filled, runs instead of Command when off is the parameter": "If filled, runs instead of Command when off is the parameter",
+          "Your command": "Your command",
+          "Trigger name": "Trigger name",
+          "Current Commands": "Current Commands"
+        }
+      },
+      pt: {
+        translation: {
+          "Add Command": "Adicionar Comando",
+          "Browse Examples": "Procurar Exemplos",
+          "Computer List": "Lista de Computadores",
+          "Operating System": "Sistema Operacional",
+          "Cancel": "Cancelar",
+          "Allow Parameters": "Permitir Parâmetros",
+          "Save": "Salvar",
+          "Alexa or Google will say this when it runs (optional)": "Alexa ou Google dirá isso quando for executado (opcional)",
+          "How to use background commands": "Como usar comandos em segundo plano",
+          "How to use Off Command": "Como usar o comando Off",
+          "Voice Reply": "Resposta de voz",
+          "Word you\'ll say to Alexa or Google (optional)": "Palavra que você dirá para Alexa ou Google (opcional)",
+          "If filled, runs instead of Command when off is the parameter": "Se preenchido, executa em vez de Command quando desligado é o parâmetro",
+          "Your command": "Seu comando",
+          "Trigger name": "Nome do gatilho",
+          "Current Commands": "Comandos Atuais"
+        }
+      }
+    }
+  });
+
+
+// import jquery from 'jquery';
+// var $ = jquery;
+// var jQuery = jquery;
+
+var $ = require('jquery');
+global.jQuery = require("jquery");
 var _ = require('lodash');
 var bootstrap = require('bootstrap');
-var fs = eRequire('fs');
 var loadApts = JSON.parse(fs.readFileSync(dataLocation));
 
 var electron = eRequire('electron');
@@ -17,10 +89,34 @@ var HeaderNav = require('./HeaderNav');
 var AddAppointment = require('./AddAppointment');
 var EditAppointment = require('./EditAppointment');
 
-var MainInterface = React.createClass({  
-  getInitialState: function() {    
-    return {
-      operatingSystem: operatingSystem,  // added here because it wouldn't detect the OS in the render function
+// var MainInterface = React.createClass({  
+class MainInterface extends React.Component {
+  constructor(props) {
+    super(props);
+    this.componentDidMount = this.componentDidMount.bind(this);
+    this.componentWillUnmount = this.componentWillUnmount.bind(this);
+    this.componentDidUpdate = this.componentDidUpdate.bind(this);
+    this.toggleAptDisplay = this.toggleAptDisplay.bind(this);
+    this.toggleEditDisplay = this.toggleEditDisplay.bind(this);
+    this.onTriggerChange = this.onTriggerChange.bind(this);
+    this.onCommandChange = this.onCommandChange.bind(this);
+    this.onOffCommandChange = this.onOffCommandChange.bind(this);
+    this.onGroundChange = this.onGroundChange.bind(this);
+    this.onVoiceChange = this.onVoiceChange.bind(this);
+    this.onVoiceReplyChange = this.onVoiceReplyChange.bind(this);
+    this.onAllowParamsChange = this.onAllowParamsChange.bind(this);
+    this.changeItem = this.changeItem.bind(this);
+    this.browseExamples = this.browseExamples.bind(this);
+    this.openComputerList = this.openComputerList.bind(this);
+    this.openGroundInstructions = this.openGroundInstructions.bind(this);
+    this.openOffCommandInstructions = this.openOffCommandInstructions.bind(this);
+    this.addItem = this.addItem.bind(this);
+    this.deleteMessage = this.deleteMessage.bind(this);
+    this.runCommand = this.runCommand.bind(this);
+    this.reOrder = this.reOrder.bind(this);
+    this.searchApts = this.searchApts.bind(this);
+    this.state = {
+      operatingSystem: '',  // added here because it wouldn't detect the OS in the render function
       aptBodyVisible: false,
       editBodyVisible: false,
       orderBy: 'trigger',
@@ -34,10 +130,29 @@ var MainInterface = React.createClass({
       editVoiceReply: '',
       editKey: null,
       myAppointments: loadApts
-    }//return
-  }, //getInitialState
+    };
+  }
 
-  componentDidMount: function() {
+  // getInitialState() {
+  //   return {
+  //     operatingSystem: '',  // added here because it wouldn't detect the OS in the render function
+  //     aptBodyVisible: false,
+  //     editBodyVisible: false,
+  //     orderBy: 'trigger',
+  //     orderDir: 'asc',
+  //     queryText: '',
+  //     editTrigger: '',
+  //     editCommand: '',
+  //     editOffCommand: '',
+  //     editGround: '',
+  //     editVoice: '',
+  //     editVoiceReply: '',
+  //     editKey: null,
+  //     myAppointments: loadApts
+  //   }//return
+  // } //getInitialState
+
+  componentDidMount() {
     ipc.on('addAppointment', function(event,message) {
       this.toggleAptDisplay();
     }.bind(this));
@@ -45,9 +160,9 @@ var MainInterface = React.createClass({
     ipc.on('editAppointment', function(event,message) {
       this.toggleEditDisplay();
     }.bind(this));
-  }, //componentDidMount
+  } //componentDidMount
 
-  componentWillUnmount: function() {
+  componentWillUnmount() {
     ipc.removeListener('addAppointment', function(event,message) {
       this.toggleAptDisplay();
     }.bind(this));
@@ -55,9 +170,9 @@ var MainInterface = React.createClass({
     ipc.removeListener('editAppointment', function(event,message) {
       this.toggleEditDisplay();
     }.bind(this));
-  }, //componentDidMount
+  } //componentDidMount
 
-  componentDidUpdate: function() {
+  componentDidUpdate() {
     if ((this.state.aptBodyVisible == false) && (this.state.editBodyVisible == false)) {
       console.log('Neither add nor edit box visible, so updating file');
 
@@ -74,17 +189,17 @@ var MainInterface = React.createClass({
         }
       });
     }
-  }, //componentDidUpdate
+  } //componentDidUpdate
 
-  toggleAptDisplay: function() {
+  toggleAptDisplay() {
     var tempVisibility = !this.state.aptBodyVisible;
     this.setState({
       aptBodyVisible: tempVisibility,
       editAllowParams: false
     }); //setState
-  }, //toggleAptDisplay
+  } //toggleAptDisplay
 
-  toggleEditDisplay: function(item) {
+  toggleEditDisplay(item) {
     var tempVisibility = !this.state.editBodyVisible;
     this.setState({
       editBodyVisible: tempVisibility,
@@ -97,45 +212,45 @@ var MainInterface = React.createClass({
       editAllowParams: item.allowParams,
       editKey: item.mykey
     }); //setState
-  }, //toggleAptDisplay
+  } //toggleAptDisplay
 
-  onTriggerChange: function(value) {
+  onTriggerChange(value) {
     this.setState({
       editTrigger: value
     }); //setState
-  },
-  onCommandChange: function(value) {
+  }
+  onCommandChange(value) {
     this.setState({
       editCommand: value
     }); //setState
-  },
-  onOffCommandChange: function(value) {
+  }
+  onOffCommandChange(value) {
     this.setState({
       editOffCommand: value
     }); //setState
-  },
-  onGroundChange: function(value) {
+  }
+  onGroundChange(value) {
     this.setState({
       editGround: value
     }); //setState
-  },
-  onVoiceChange: function(value) {
+  }
+  onVoiceChange(value) {
     this.setState({
       editVoice: value
     }); //setState
-  },
-  onVoiceReplyChange: function(value) {
+  }
+  onVoiceReplyChange(value) {
     this.setState({
       editVoiceReply: value
     }); //setState
-  },
-  onAllowParamsChange: function(value) {
+  }
+  onAllowParamsChange(value) {
     this.setState({
       editAllowParams: value
     }); //setState
-  },
+  }
 
-  changeItem: function(item) {
+  changeItem(item) {
     var allApts = this.state.myAppointments;
     var newApts = _.without(allApts, this.state.myAppointments[item.mykey]);
 
@@ -158,60 +273,60 @@ var MainInterface = React.createClass({
       myAppointments: newApts,
       aptBodyVisible: false
     }) //setState
-  }, //changeItem
+  } //changeItem
 
-  browseExamples: function() {
+  browseExamples(e) {
     ipc.sendSync('openexampleWindow');
-  }, //browseExamples
+  } //browseExamples
 
-  openComputerList: function() {
+  openComputerList(e) {
     electron.shell.openExternal('https://www.triggercmd.com/user/computer/list');
-  }, //openComputerList
+  } //openComputerList
 
-  openGroundInstructions: function() {
+  openGroundInstructions(e) {
     electron.shell.openExternal('https://www.triggercmd.com/forum/topic/15/what-s-the-difference-between-background-and-foreground-commands');
-  }, 
+  } 
 
-  openOffCommandInstructions: function() {
+  openOffCommandInstructions(e) {
     electron.shell.openExternal('https://www.triggercmd.com/forum/topic/853/how-to-use-off-command');
-  }, 
+  } 
 
-  addItem: function(tempItem) {    
+  addItem(tempItem) {
     var tempApts = this.state.myAppointments;
     tempApts.push(tempItem);
     this.setState({
       myAppointments: tempApts,
       aptBodyVisible: false
     }) //setState
-  }, //addItem
+  } //addItem
 
-  deleteMessage: function(item) {
+  deleteMessage(item) {
     var allApts = this.state.myAppointments;
     var newApts = _.without(allApts, item);
     this.setState({
       myAppointments: newApts
     }); //setState
-  }, //deleteMessage
+  } //deleteMessage
 
-  runCommand: function(item) {
+  runCommand(item) {
     console.log('Running ' + item.command);
     var ChildProcess = cp.exec(item.command);
-  }, //runCommand
+  } //runCommand
 
-  reOrder: function(orderBy, orderDir) {
+  reOrder(orderBy, orderDir) {
     this.setState({
       orderBy: orderBy,
       orderDir: orderDir
     }) //setState
-  }, //reOrder
+  } //reOrder
 
-  searchApts: function(query) {
+  searchApts(query) {
     this.setState({
       queryText: query
     }); //setState
-  }, //searchApts
+  } //searchApts
 
-  render: function() {
+  render() {
     var filteredApts = [];
     var queryText = this.state.queryText;
     var orderBy = this.state.orderBy;
@@ -231,10 +346,9 @@ var MainInterface = React.createClass({
     }
     for (var i = 0; i < myAppointments.length; i++) {
       if (
-        (myAppointments[i].trigger.toLowerCase().indexOf(queryText)!=-1) ||
-        (myAppointments[i].command.toLowerCase().indexOf(queryText)!=-1) ||
-        (myAppointments[i].ground.toLowerCase().indexOf(queryText)!=-1) ||
-        (myAppointments[i].voice.toLowerCase().indexOf(queryText)!=-1)
+        (myAppointments[i].trigger && myAppointments[i].trigger.toLowerCase().indexOf(queryText)!=-1) ||
+        (myAppointments[i].command && myAppointments[i].command.toLowerCase().indexOf(queryText)!=-1) ||
+        (myAppointments[i].voice && myAppointments[i].voice.toLowerCase().indexOf(queryText)!=-1)
 
       ) {
         filteredApts.push(myAppointments[i]);
@@ -322,7 +436,7 @@ var MainInterface = React.createClass({
           <div className="container">
            <div className="row">
              <div className="appointments col-sm-12">
-               <h2 className="appointments-headline">Current Commands</h2>
+               <h2 className="appointments-headline">{i18next.t('Current Commands')}</h2>
                <ul className="item-list media-list">{filteredApts}</ul>
              </div>{/* col-sm-12 */}
            </div>{/* row */}
@@ -331,7 +445,7 @@ var MainInterface = React.createClass({
       </div>
     );
   } //render
-});//MainInterface
+};//MainInterface
 
 ReactDOM.render(
   <MainInterface />,
